+++ packages/ConsoleDiffer/src/DifferAndFormatter.php
@@ -17,16 +17,10 @@ final class DifferAndFormatter
      */
     private $diffConsoleFormatter;
 
-    /**
-     * @var Differ
-     */
-    private $bareDiffer;
-
-    public function __construct(Differ $differ, DiffConsoleFormatter $diffConsoleFormatter, Differ $bareDiffer)
+    public function __construct(Differ $differ, DiffConsoleFormatter $diffConsoleFormatter)
     {
         $this->differ = $differ;
         $this->diffConsoleFormatter = $diffConsoleFormatter;
-        $this->bareDiffer = $bareDiffer;
     }
 
     public function diffAndFormat(string $old, string $new): string
@@ -38,41 +32,5 @@ final class DifferAndFormatter
         $diff = $this->differ->diff($old, $new);
 
         return $this->diffConsoleFormatter->format($diff);
-    }
-
-    /**
-     * Returns only the diff (- and + lines), no extra elements, lines nor ----- around it
-     */
-    public function bareDiffAndFormat(string $old, string $new): string
-    {
-        $diff = $this->bareDiffAndFormatWithoutColors($old, $new);
-        if ($diff === '') {
-            return '';
-        }
-
-        return $this->diffConsoleFormatter->bareFormat($diff);
-    }
-
-    public function bareDiffAndFormatWithoutColors(string $old, string $new): string
-    {
-        if ($old === $new) {
-            return '';
-        }
-
-        $diff = $this->bareDiffer->diff($old, $new);
-
-        $diff = $this->removeTrailingWhitespaces($diff);
-
-        // impossible to configure - removed manually
-        return substr($diff, strlen('@@ @@ '));
-    }
-
-    private function removeTrailingWhitespaces(string $diff): string
-    {
-        $diff = preg_replace('#\n( )\n#', PHP_EOL . PHP_EOL, $diff);
-
-        $diff = preg_replace('#( )\n#', PHP_EOL, $diff);
-
-        return rtrim($diff);
     }
 }
