+++ packages/ConsoleDiffer/src/config/external-services.yml
@@ -12,13 +12,13 @@ services:
         arguments:
             - '@diffOutputBuilder'
 
-    # bare differ
-    bareDiffOutputBuilder:
+    # makrdown differ
+    markdownDiffOutputBuilder:
         class: SebastianBergmann\Diff\Output\UnifiedDiffOutputBuilder
         arguments:
             $header: ''
 
-    bareDiffer:
+    markdownDiffer:
         class: SebastianBergmann\Diff\Differ
         arguments:
-            - '@bareDiffOutputBuilder'
+            - '@markdownDiffOutputBuilder'
