+++ packages/ConsoleDiffer/src/config/services.yml
@@ -10,4 +10,6 @@ services:
 
     Rector\ConsoleDiffer\DifferAndFormatter:
         $differ: '@differ'
-        $bareDiffer: '@bareDiffer'
+
+    Rector\ConsoleDiffer\MarkdownDifferAndFormatter:
+        $markdownDiffer: '@markdownDiffer'
